requirements.txt
pyTelegramBotAPI==4.26.0
psutil==6.1.1
python-dotenv==1.0.1
matplotlib==3.10.0
speedtest-cli==2.1.3
requests==2.32.3
gpiozero==2.0.1
